Apply envvar from process template in alignprocess
This looks like a really big change, but it's pretty simple. Instead of
writing out our env vars immediately to the file; we store them in an
OrderedDict (just like a regular dict, except it preserves insert
order). We then layer on top the changes specified by the process
template variables field. Once we're done, we write that to the
outfile. It would probably be easy to factor the environment variable
determination out into a separate function, but I haven't done that
here.

Special note: Setting a variable to None (or null in the JSON) will
cause that variable to be unset, rather than set to any other particular
value.