Fix rare Picard MarkDups failure
Explanation:
Picard 1.118 contains a bug in which it incorrectly processed 'new'
format Illumina read names. This bug is fixed in 1.119 forwards.

These leads to the Optical Duplicate routine having to deal with huge
chunks of data - causing it to sometimes take a very long time, and in
one rare case seen today; crash entirely (probably due to another bug).

This fix manually specifies the READ_NAME_REGEX rather than rely on the
default splitting behavior, which should:
1) prevent crashes
2) increase accuracy
3) potentially decrease runtime

For more details, see this mailing list thread:
http://sourceforge.net/p/samtools/mailman/samtools-help/thread/53F81EE6.30401@uw.edu/
and this commit:
https://github.com/broadinstitute/picard/commit/3fcb0fc