star: Disable LoadAndStore directive
This feature was occasionally causing processes to wait indefinitely.  This
occurred when one process took responsibility for loading the genome into
memory, failed to do so, and never relinquishes responsibility. Because we have
a lot of RAM and we're already reserving it for each process, it is easier to
disable this feature than to work around it at this time.