Autoaggregation - proceed without failed alignments
With the rate of failures that have been occurring recently, I think it
makes sense to complete aggregations on a best-effort basis.

This will launch all aggregations that it can once all alignments have
finished, successfully or otherwise.

It will make relaunching aggregations a little more tricky, but I think
the tradeoff is worth it.