Fix malformed test data & update results
The input files `r1.fastq.gz` and `r2.fastq.gz` contained accidental
lines like `--`, which is invalid for a fastq format. This caused a few
reads to be discarded when splitting fastq files, and the `pf` and
`total` counts to be notably different in `tagcounts.txt`.

The effect on the overall results was low but scattered, and the
expected test result files have been updated to be correct for the new
input files.

This error likely stemmed from initial creation of the test fastq files,
as they resemble grep line separators.